feat(plasma-b2c/web): separate configs for segment
